remove rounding error

+++ app/src/main/java/tud/tk3/ambientdisplay/app/communication/DumbDisplayController.java
@@ -51,15 +51,15 @@ public class DumbDisplayController implements DisplayController {
         // Calculate total width
         for(int i = sorted.size() - 1; i >= 0; i--){
             display = sorted.get(i);
-            totalX += display.width / display.dpi;
+            totalX += (double)display.width / display.dpi;
             //totalY = Math.max(totalY,display.height / display.dpi);
         }
         totalY = totalX;
         // Generate configuration
         for(int i = sorted.size() - 1; i >= 0; i--){
             display = sorted.get(i);
-            offsetX = (display.width / (display.dpi * totalX));
-            offsetY = display.height / (display.dpi * totalY);
+            offsetX = (double)display.width / (display.dpi * totalX);
+            offsetY = (double)display.height / (display.dpi * totalY);
             //offsetY = offsetX * (display.height/display.width); //maximize screen estate, dirty hack ;)
             //offsetY = display.height / (display.dpi * totalX);
             ImageSection section = new ImageSection(posX, posY, offsetX, offsetY);
